requirements.txt
flask
opencv-python-headless
numpy
requests
pillow
piexif
gunicorn
